requirements.txt
# Core Data Processing
pandas==2.1.3
numpy==1.26.4
psycopg2-binary==2.9.9

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
imbalanced-learn==0.11.0

# Data Validation
great-expectations==0.18.8

# Database
sqlalchemy==2.0.23

# ML Tracking
mlflow==2.9.2

# API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0
streamlit==1.29.0

# Network Analysis
networkx==3.2.1

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1